backend/requirements.txt
asgiref==3.7.2
Brotli==1.0.9
cairocffi==1.6.1
CairoSVG==2.7.0
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
cssselect2==0.7.0
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.3
django-colorfield==0.9.0
django-cors-headers==4.1.0
django-debug-toolbar==3.2.4
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-extra-fields==3.6.1
filelock==3.12.2
filetype==1.2.0
flake8==6.0.0
flake8-import-order==0.18.2
fonttools==4.41.1
html5lib==1.1
identify==2.5.26
idna==3.4
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
nodeenv==1.8.0
oauthlib==3.2.2
Pillow==10.0.0
platformdirs==3.9.1
pre-commit==3.3.3
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pydyf==0.7.0
pyflakes==3.0.1
PyJWT==2.7.0
pyphen==0.14.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
tinycss2==1.2.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.3
virtualenv==20.24.2
WeasyPrint==52.5
webcolors==1.13
webencodings==0.5.1
zopfli==0.2.2
gunicorn==20.1.0